+++ src/globals.js
@@ -64,8 +64,11 @@ const getLoggingLevel = () => {
 var qrsUtil = require('./qrsUtil');
 
 // Helper function to read the contents of the certificate files:
-// const readCert = filename => fs.readFileSync(path.resolve(__dirname, certificatesPath, filename));
 const readCert = filename => fs.readFileSync(filename);
+
+var certPath = path.resolve(__dirname, config.get('Butler.cert.clientCert')),
+    keyPath = path.resolve(__dirname, config.get('Butler.cert.clientCertKey')),
+    caPath = path.resolve(__dirname, config.get('Butler.cert.clientCertCA'));
 
 //  Engine config
 var configEngine = {
@@ -87,9 +90,14 @@ const configQRS = {
     useSSL: config.get('Butler.configQRS.useSSL'),
     headerKey: config.get('Butler.configQRS.headerKey'),
     headerValue: config.get('Butler.configQRS.headerValue'),
-    cert: readCert(config.get('Butler.cert.clientCert')),
-    key: readCert(config.get('Butler.cert.clientCertKey')),
-    ca: readCert(config.get('Butler.cert.clientCertCA')),
+    cert: readCert(certPath),
+    key: readCert(keyPath),
+    ca: readCert(caPath),
+    certPaths: {
+        certPath: certPath,
+        keyPath: keyPath,
+        capath: caPath,
+    },
 };
 
 // ------------------------------------
@@ -108,7 +116,6 @@ if (config.has('Butler.teamsConfig.enable') && config.has('Butler.teamsConfig.ta
 
     // Create MS Teams object
     var teamsTaskFailureObj = new IncomingWebhook(teamsTaskFailureURL);
-
 }
 
 // ------------------------------------
@@ -171,7 +178,6 @@ if (config.has('Butler.fileCopyApprovedDirectories')) {
         fileCopyDirectories.push(newDirCombo);
     });
 }
-
 
 // Load approved fromDir and toDir for fileMove operation
 var fileMoveDirectories = [];